+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: brunofer <brunofer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/14 07:02:52 by brunofer          #+#    #+#             */
-/*   Updated: 2025/07/15 11:06:52 by brunofer         ###   ########.fr       */
+/*   Updated: 2025/07/15 13:32:20 by brunofer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t n)
 {
 	char			number;
-	int				i;
+	size_t			i;
 	unsigned char	*ptr;
 
 	number = (unsigned char)c;
